libtool: add -os2dllname option
OS/2 limits a length of a DLL base name up to 8 characters. Because of
this, if there are many DLLs whose prefix is same and very long, the
result DLLs are overwritten. So to avoid this, the option to specify a
OS/2 DLL name is needed.

* NEWS: Add news for -os2dllname.
* doc/libtool.texi: Add -os2dllname item.
* libltdl/config/ltmain.m4sh (func_mode_help): Add a description for
-os2dllname.
(fund_mode_link): Add -os2dllname.